requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
requests
beautifulsoup4
lxml
passlib[bcrypt]
python-jose[cryptography]
redis
pydantic[email]
python-multipart
pytest
pytest-cov
httpx
pytest-mock
